+++ src/region/wrapConnect.js
@@ -6,13 +6,9 @@ const Empty = () => null;
 
 export default (Region) => {
   class RegionConnect extends Region {
-    connectWith = (key, Display, option = {}) => {
+    connectWith = (key, Display, option) => {
       const { connect } = this;
-      if (typeof option === 'object') {
-        return connect(key, option)(Display);
-      }
-      console.warn('connectWith receives a wide option, the original param is deprecated, replace with connectWith(key, Display { Loading: LoadingComponent })');
-      return connect(key, { Loading: option })(Display);
+      return connect(key, option)(Display);
     }
 
     connect = (key, { Loading, Error } = {}) => (Display = Empty) => {